Make Bars stack by default and remove StackedBars
