+++ packages/visualizations/src/index.ts
@@ -1,7 +1,6 @@
 export { Axis } from "./Axis";
 export { Area } from "./Area";
 export { Bars } from "./Bars";
-export { StackedBars } from "./StackedBars";
 export { Line } from "./Line";
 export { Chart, ChartProps } from "./Chart";
 export { AxialChart, AxialChartProps } from "./types";
